requirements.txt
pandas>=2.0
numpy>=1.24
chromadb>=0.4.22
sentence-transformers>=2.2.2
flask>=2.3
requests>=2.31
pyinstaller>=6.0
python-dotenv>=1.0.0
